Revert "[CSOptimizer] Initial implementation of disjunction choice favoring algorithm"
This reverts commit 672ae3d252b6030c83a179fcc2e67704c0a758d8.
