+++ validation-test/Sema/type_checker_perf/slow/rdar22770433.swift
@@ -2,6 +2,7 @@
 // REQUIRES: tools-release,no_asan
 
 func test(n: Int) -> Int {
+  // expected-error@+1 {{the compiler is unable to type-check this expression in reasonable time}}
   return n == 0 ? 0 : (0..<n).reduce(0) {
     ($0 > 0 && $1 % 2 == 0) ? ((($0 + $1) - ($0 + $1)) / ($1 - $0)) + (($0 + $1) / ($1 - $0)) : $0
   }
