+++ validation-test/Sema/type_checker_perf/slow/rdar31742586.swift
@@ -3,4 +3,5 @@
 
 func rdar31742586() -> Double {
   return -(1 + 2) + -(3 + 4) + 5 - (-(1 + 2) + -(3 + 4) + 5)
+  // expected-error@-1 {{the compiler is unable to type-check this expression in reasonable time}}
 }
