+++ validation-test/Sema/type_checker_perf/slow/rdar46713933_literal_arg.swift
@@ -8,4 +8,5 @@ func wrap<T: ExpressibleByStringLiteral>(_ key: String, _ value: T) -> T { retur
 
 func wrapped() -> Int {
   return wrap("1", 1) + wrap("1", 1) + wrap("1", 1) + wrap("1", 1) + wrap("1", 1) + wrap("1", 1)
+  // expected-error@-1 {{the compiler is unable to type-check this expression in reasonable time}}
 }
