+++ validation-test/Sema/type_checker_perf/slow/rdar91310777.swift
@@ -9,6 +9,7 @@ func test() {
     compute {
       print(x)
       let v: UInt64 = UInt64((24 / UInt32(1)) + UInt32(0) - UInt32(0) - 24 / 42 - 42)
+      // expected-error@-1 {{the compiler is unable to type-check this expression in reasonable time; try breaking up the expression into distinct sub-expressions}}
       print(v)
     }
   }
